Deprecate Menu templates in favor of Ryzom Components
This is a work in progress that requires the ssr branch from Ryzom to
work.

This shows how dealing HTML from Python data could be more enjoyable to
some users with this pattern of encapsulation, than with a template
layer that was made to reduce programing power of HTML scripting.

We can see two different kind of syntaxes, the initial one that comes
from JS/React, and a pythonic one that takes *content, **attrs on
Component classes.

The performance benefit should be notable because we removed two
template files that were used in base.html

This does not demonstrate channels based queryset data binding because
the menu is only a part of the layout. Also, it seems that we will have
to wait until all our templates have been removed in favor of Ryzom
components before we can remove Turbolinks and enable Ryzom isomorphic
features such as component events, client side rendering and channels
data-binding.