wip: HTML fixed delay live reload with the poll attribute
Problem is: the triggers the unpoly "Navigation feedback" which I have
recently implemented ... with a full screen overlay!!! Damned, other
than this blocking UI problem it's working as expected.

So, I tried to make the loader smaller and without an overlay to fit in
the top-right corner of the screen where we typically have nothing and
plan on implementing a "search anything" input bar.

Original idea and code from:

https://github.com/unpoly/unpoly/issues/63

But changed from setTimer to recursive setTimeout to unsure a fixed
delay between each code (set in the HTML poll attribute).

Also, converted from JS to Python which makes it much more satisfying to
read and write in my opinion /troll

Related Unpoly documentation:

https://unpoly.com/up.reload
https://unpoly.com/up.feedback